Include static index file in bundle
