+++ src/run/bundle.rs
@@ -5,11 +5,16 @@ use std::{
 
 use super::{cargo::metadata::Metadata, BinTarget};
 
+pub enum Index {
+    File(PathBuf),
+    Static(&'static str),
+}
+
 pub struct LinkedBundle {
     wasm_path: PathBuf,
     js_path: PathBuf,
     assets_path: Option<PathBuf>,
-    custom_index: Option<PathBuf>,
+    index: Index,
 }
 
 pub struct PackedBundle {
@@ -42,7 +47,7 @@ pub fn create_bundle(
             None
         },
         // TODO: Determine if index is customized
-        custom_index: None,
+        index: Index::Static(default_index(&bin_target)),
     };
 
     if packed {
@@ -68,4 +73,24 @@ pub fn create_bundle(
     } else {
         Ok(WebBundle::Linked(linked))
     }
+}
+
+/// Create the default `index.html` if the user didn't provide one.
+pub fn default_index(bin_target: &BinTarget) -> &'static str {
+    let template = include_str!(concat!(
+        env!("CARGO_MANIFEST_DIR"),
+        "/assets/web/index.html"
+    ));
+
+    // Insert correct path to JS bindings
+    let index = template.replace(
+        "./build/bevy_app.js",
+        format!("./build/{}.js", bin_target.bin_name).as_str(),
+    );
+
+    // Only static strings can be served in the web app,
+    // so we leak the string memory to convert it to a static reference.
+    // PERF: This is assumed to be used only once and is needed for the rest of the app running
+    // time, making the memory leak acceptable.
+    Box::leak(index.into_boxed_str())
 }
